Add ability to sendPrivate
The IRC adapter added `sendPrivate` for redirecting long output to privmsg. Making this the default on IRC would be amazing. 

IDK if people want to add something to configure it to send to channel, but in most cases, I don't think people want all the output going to the whole IRC channel by default.