Add cross-package 'help:lookup' listener
Problem: You create a script package that has a namespaced command, such as:

```
user>> hubot widget me foobar
hubot>> Your foobar widget is coming right up!
```

Inevitably, it gets more complex than that, so you are tempted to add something like this:

```
robot.respond /widget help/, (msg) ->
  msg.send "#{robot.name} widget me <something> - get a lovely widget"
  [...]
```

This is lame, because you already have the command list in the header. And for some reason, folks can never remember to put the `help` before the namespaced command.

This PR offers developers the ability to add a boilerplate:

```
robot.respond /widget help/, (msg) ->
  robot.emit 'help:lookup', msg
```

and it will route back through the commonly-installed `hubot-help` package and reflect back its help text.

---

This is a proposed idea -- there may be a better way to do it using Hubot internals (and without changing this package), but I could not find one.