Add LICENSE
The name of the license under which the script is supposed to be
licensed and the name of the original author were taken from
package.json. I'm happy that the script is licensed under the MIT
License, however the sources of the script should be accompanied by the
text of the license, otherwise it might appear that the script is
proprietary.